requirements.txt
asgiref==3.8.1
Django==5.0.6
django-debug-toolbar==4.3.0
django-environ==0.11.2
django-rest-framework==0.1.0
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
pillow==10.3.0
psycopg2==2.9.9
PyJWT==2.8.0
sqlparse==0.5.0
typing_extensions==4.11.0
tzdata==2024.1
